+++ dbt/dbt_company/models/staging/staging_transactions_resellers_xml.sql
@@ -60,11 +60,11 @@ SELECT
     date_bought,
     'YYYYMMDD'
   ) :: INT AS bought_date_key,
-  total_amount,
+  total_amount::numeric,
   no_purchased_postcards,
-  e.product_price,
+  e.product_price::numeric,
   e.geography_key,
-  s.commission_pct * total_amount AS commisionpaid,
+  s.commission_pct * total_amount::numeric AS commisionpaid,
   s.commission_pct,
   loaded_timestamp
 FROM
